scripts/restart_workflows.py
#!/usr/bin/env python3
"""
Script to restart failed PyTorch workflows.
Queries PyTorch HUD and ClickHouse to identify and restart failed workflows.
"""

import os
import sys
import logging
from dotenv import load_dotenv

# Client libraries
# from pytorch_hud_client import HUDClient  # TODO: Import from claude-pytorch-treehugger
# from clickhouse_mcp import ClickHouseClient  # TODO: Import from clickhouse-mcp

def setup_logging():
    """Configure logging based on environment variables."""
    log_level = os.getenv('LOG_LEVEL', 'INFO').upper()
    logging.basicConfig(
        level=getattr(logging, log_level),
        format='%(asctime)s - %(levelname)s - %(message)s'
    )
    return logging.getLogger(__name__)

def load_config():
    """Load configuration from environment variables."""
    load_dotenv()
    
    config = {
        'pytorch_hud_url': os.getenv('PYTORCH_HUD_URL'),
        'pytorch_hud_token': os.getenv('PYTORCH_HUD_TOKEN'),
        'clickhouse_host': os.getenv('CLICKHOUSE_HOST'),
        'clickhouse_port': int(os.getenv('CLICKHOUSE_PORT', 9000)),
        'clickhouse_user': os.getenv('CLICKHOUSE_USER', 'default'),
        'clickhouse_password': os.getenv('CLICKHOUSE_PASSWORD'),
        'clickhouse_database': os.getenv('CLICKHOUSE_DATABASE', 'pytorch'),
        'dry_run': os.getenv('DRY_RUN', 'false').lower() == 'true'
    }
    
    # Validate required config
    required_fields = ['pytorch_hud_url', 'pytorch_hud_token', 'clickhouse_host']
    missing_fields = [field for field in required_fields if not config[field]]
    
    if missing_fields:
        raise ValueError(f"Missing required environment variables: {', '.join(missing_fields)}")
    
    return config

def query_pytorch_hud(config, logger):
    """Query PyTorch HUD for failed workflows."""
    logger.info("Querying PyTorch HUD for failed workflows...")
    # TODO: Initialize HUD client and query for failed workflows
    # hud_client = HUDClient(config['pytorch_hud_url'], config['pytorch_hud_token'])
    # failed_workflows = hud_client.get_failed_workflows()
    return []

def query_clickhouse(config, logger):
    """Query ClickHouse for workflow history."""
    logger.info("Querying ClickHouse for workflow history...")
    # TODO: Initialize ClickHouse client and query workflow history
    # ch_client = ClickHouseClient(
    #     host=config['clickhouse_host'],
    #     port=config['clickhouse_port'], 
    #     user=config['clickhouse_user'],
    #     password=config['clickhouse_password'],
    #     database=config['clickhouse_database']
    # )
    # workflow_history = ch_client.query("SELECT * FROM workflows WHERE status = 'failed'")
    return []

def restart_workflows(failed_workflows, config, logger):
    """Restart the identified failed workflows."""
    if config['dry_run']:
        logger.info(f"DRY RUN: Would restart {len(failed_workflows)} workflows")
        return
    
    logger.info(f"Restarting {len(failed_workflows)} workflows...")
    # TODO: Implement workflow restart logic

def main():
    """Main entry point."""
    try:
        logger = setup_logging()
        config = load_config()
        
        logger.info("Starting workflow restart process...")
        
        # Query for failed workflows
        failed_workflows = query_pytorch_hud(config, logger)
        
        # Get additional context from ClickHouse
        workflow_history = query_clickhouse(config, logger)
        
        # Restart workflows
        restart_workflows(failed_workflows, config, logger)
        
        logger.info("Workflow restart process completed successfully")
        
    except Exception as e:
        logger.error(f"Error during workflow restart: {e}")
        sys.exit(1)

if __name__ == "__main__":
    main()